refactor(dynacell): home the shared window tiler in dynacell.tiling
window_starts is model-agnostic but lived in celldiff_wrapper, so the
generic engine imported a tiler from a CELLDiff-specific module. Move it
to dynacell.tiling, which both consumers can import without engine ->
celldiff_wrapper cycling.

Two simplifications ride along, since the function relocated wholesale:

- the start-index loop collapses to [*range(0, last, stride), last].
  Verified equivalent to the while-loop over every (size, patch, overlap)
  for patch 1..32 and extents up to patch+199 (105,600 triples, 0
  mismatches), including size == patch where last == 0 yields [0].
- _as_overlap_triple is replaced by monai.utils.ensure_tuple_rep, called
  inside window_starts rather than at 2 of the 4 call sites. Checked
  identical on int, tuple, list and all four rank-error inputs. Callers
  now pass overlap_size straight through, and a bare int broadcasts for
  every caller instead of only the two that normalized first.

Error behavior is unchanged: patch-larger-than-extent, overlap outside
[0, patch), and rank mismatch all still raise ValueError.

(cherry picked from commit 932c649ccd9cd54d76a96ba83d5a9a3a1b939887)
